Started to add environmental config.
