+++ README.md
@@ -9,6 +9,6 @@ In your IDE run the file 'DefaultRunner.java'.
 # Run Tests in Parallel
 
 Using a command line tool, navigate to the root directory then run the command 'mvn clean install'.
-This will run 4 runners in parallel.  Details of the parallel execution can be found in the 'pom.xml' file. 
+This will run 4 com.automatedtesting.pages.runners in parallel.  Details of the parallel execution can be found in the 'pom.xml' file.
