+++ pom.xml
@@ -14,6 +14,7 @@
         <hamcrest.version>1.3</hamcrest.version>
         <selenium.version>3.11.0</selenium.version>
         <log4j.version>1.2.17</log4j.version>
+        <spring-boot.version>2.1.5.RELEASE</spring-boot.version>
         <maven-failsafe-plugin.version>3.0.0-M3</maven-failsafe-plugin.version>
     </properties>
 
@@ -38,7 +39,19 @@
         </dependency>
         <dependency>
             <groupId>io.cucumber</groupId>
+            <artifactId>cucumber-java8</artifactId>
+            <version>${cucumber.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.cucumber</groupId>
             <artifactId>cucumber-junit</artifactId>
+            <version>${cucumber.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.cucumber</groupId>
+            <artifactId>cucumber-spring</artifactId>
             <version>${cucumber.version}</version>
             <scope>test</scope>
         </dependency>
@@ -60,6 +73,13 @@
             <version>${log4j.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <version>${spring-boot.version}</version>
+            <scope>test</scope>
+        </dependency>
+
     </dependencies>
 
     <build>
@@ -85,7 +105,7 @@
                         <!--<include>**/Default.java</include>-->
 
                         <!-- UNCOMMENT BELOW LINE - To execute feature files with multiple
-                            runners -->
+                            com.automatedtesting.pages.runners -->
                          <include>**/Runner*.java</include>
                     </includes>
