+++ src/test/java/com/automatedtesting/pages/BasePage.java
@@ -1,0 +1,7 @@
+package com.automatedtesting.pages;
+
+import org.openqa.selenium.WebDriver;
+
+public abstract class BasePage {
+    protected abstract void goToPage();
+}
