+++ src/test/java/com/automatedtesting/steps/BasePageSteps.java
@@ -1,0 +1,8 @@
+package com.automatedtesting.steps;
+
+import org.openqa.selenium.WebDriver;
+import com.automatedtesting.support.DriverSupport;
+
+class BasePageSteps {
+    WebDriver driver = new DriverSupport().getDriver();
+}
