+++ src/test/java/com/automatedtesting/support/DriverSupport.java
@@ -1,0 +1,40 @@
+package com.automatedtesting.support;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+import java.util.concurrent.TimeUnit;
+
+public class DriverSupport {
+
+    private static WebDriver driver;
+
+    public WebDriver getDriver() {
+        if (driver == null) {
+            System.setProperty("webdriver.chrome.driver", "src/test/resources/webdrivers/windows/chromedriver");
+            driver = new ChromeDriver(getChromeOptions());
+            driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+        }
+
+        return driver;
+    }
+
+    private ChromeOptions getChromeOptions() {
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("disable-infobars");
+        return options;
+    }
+
+    private static final Thread CLOSE_THREAD = new Thread() {
+        @Override
+        public void run() {
+            driver.close();
+        }
+    };
+
+    static {
+        Runtime.getRuntime().addShutdownHook(CLOSE_THREAD);
+    }
+
+}
