--- src/test/java/runners/Runner3.java
@@ -1,17 +1,0 @@
-package runners;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.SnippetType;
-import cucumber.api.junit.Cucumber;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        tags = "@runner3",
-        features = "src/test/resources/features",
-        glue = { "steps" },
-        snippets = SnippetType.CAMELCASE,
-        plugin = {"pretty", "html:target/cucumber-reports/html"})
-
-public class Runner3 {
-}
