--- src/test/java/steps/BasePageSteps.java
@@ -1,8 +1,0 @@
-package steps;
-
-import org.openqa.selenium.WebDriver;
-import support.DriverSupport;
-
-class BasePageSteps {
-    WebDriver driver = new DriverSupport().getDriver();
-}
