+++ src/test/resources/config/application.properties
@@ -1,0 +1,5 @@
+uri: http://www.math.com/students/calculators/source/
+web_drivers_path: src/test/resources/webdrivers/
+mac_driver_path: mac/chromedriver
+windows_driver_path: windows/chromedriver
+driver: webdriver.chrome.driver
