--- src/test/resources/config/common.properties
@@ -1,6 +1,0 @@
-base_uri: http://www.math.com/students/calculators/source/
-web_drivers_path: src/test/resources/webdrivers/
-linux_driver_path: linux/chromedriver
-mac_driver_path: mac/chromedriver
-windows_driver_path: windows/chromedriver 
-driver: webdriver.chrome.driver
